Gate payjoin-ffi bitcoind types behind feature
Split _test-utils into _test-utils (constants + TestServices only)
and _test-utils-bitcoind (adds BitcoindEnv, RpcClient, etc.).
This lets FFI consumers like Dart unit tests import test constants
without triggering the corepc-node/bitcoind download.

Update all binding-generation and test scripts to use
_test-utils-bitcoind where integration-test types are needed.
