+++ payjoin-ffi/Cargo.toml
@@ -9,7 +9,8 @@ exclude = ["tests"]
 default = []
 csharp = ["dep:uniffi-bindgen-cs"]
 dart = ["dep:uniffi-dart"]
-_test-utils = ["payjoin-test-utils", "payjoin-test-utils/bitcoind", "tokio"]
+_test-utils = ["payjoin-test-utils", "tokio"]
+_test-utils-bitcoind = ["_test-utils", "payjoin-test-utils/bitcoind"]
 _manual-tls = ["payjoin/_manual-tls"]
 wasm_js = ["getrandom/js"]
