+++ payjoin-ffi/contrib/test.sh
@@ -4,7 +4,7 @@ set -e
 RUST_VERSION=$(rustc --version | awk '{print $2}')
 
 if [[ ! $RUST_VERSION =~ ^1\.85\. ]]; then
-    cargo test --package payjoin-ffi --verbose --features=_manual-tls,_test-utils
+    cargo test --package payjoin-ffi --verbose --features=_manual-tls,_test-utils-bitcoind
 else
     echo "Skipping payjoin-ffi tests for Rust version $RUST_VERSION (MSRV)"
 fi
